+++ site/Generic/ExceptionX.cs
@@ -1,23 +1,23 @@
 ﻿using System;
 using System.IO;
-using Newtonsoft.Json;
 
 namespace DFM.Generic
 {
 	public static class ExceptionX
 	{
-		public static void TryLog(this Exception exception)
+		public static void TryLog(this ErrorLog error)
 		{
 			try
 			{
-				var date = DateTime.Now.ToString("yyyyMMddHHmmssffffff");
-				var path = String.Format(Cfg.LogPathErrors, date);
-
-				var json = JsonConvert.SerializeObject(exception, Formatting.Indented);
-
-				File.WriteAllText(path, json);
+				var path = String.Format(Cfg.LogPathErrors, error.ID);
+				File.WriteAllText(path, error.ToString());
 			}
 			catch { /* ignored, nothing can be done anymore */ }
+		}
+
+		public static void TryLog(this Exception exception)
+		{
+			new ErrorLog(exception).TryLog();
 		}
 	}
 }