+++ docs/RELEASES.md
@@ -21,7 +21,7 @@ Legend:
 - [ ] Search by move/detail
 
 ## <a name="4.2.0.2"></a>4.2.0.2 :sheep: <sup>`1`</sup>
-- [ ] Send site errors to cellphone 
+- [ ] `200827>......` Send site errors to cellphone 
 
 ## <a name="4.2.0.1"></a>4.2.0.1 :sheep: <sup>`8`</sup>
 - [x] `200825>200826` Add different icon for test at android